change session
