+++ config/default.conf
@@ -33,7 +33,8 @@ server {
     }
 
     location /cookie {
-        root www/html/cookie;
+        root ;
+        upload_store cookie/session;
         # index cgi-bin.html;
         index session.py;
         allowed_methods GET POST;
