+++ cookie/session.py
@@ -1,0 +1,382 @@
+import sys
+
+file_path = sys.argv[1]
+
+
+html_content = """
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>42 Webserv Space Station</title>
+    <style>
+        * {
+            margin: 0;
+            padding: 0;
+            box-sizing: border-box;
+        }
+
+        body {
+            font-family: 'Arial', sans-serif;
+            background: linear-gradient(to bottom, #0B1026 0%, #1B2735 50%, #090A0F 100%);
+            color: #fff;
+            min-height: 100vh;
+            padding: 20px;
+            position: relative;
+            overflow-x: hidden;
+        }
+
+        body::before {
+            content: '';
+            position: absolute;
+            width: 100%;
+            height: 100%;
+            top: 0;
+            left: 0;
+            background-image: 
+                radial-gradient(2px 2px at 20px 30px, #ffffff, rgba(0,0,0,0)),
+                radial-gradient(2px 2px at 40px 70px, #ffffff, rgba(0,0,0,0)),
+                radial-gradient(2px 2px at 50px 160px, #ffffff, rgba(0,0,0,0)),
+                radial-gradient(2px 2px at 90px 40px, #ffffff, rgba(0,0,0,0)),
+                radial-gradient(2px 2px at 130px 80px, #ffffff, rgba(0,0,0,0));
+            background-repeat: repeat;
+            animation: twinkle 5s infinite linear;
+            z-index: -1;
+        }
+
+        @keyframes twinkle {
+            from { transform: translateY(0); }
+            to { transform: translateY(-100px); }
+        }
+
+        .container {
+            max-width: 800px;
+            margin: 0 auto;
+            position: relative;
+        }
+
+        p {
+            margin-bottom: 10px;
+        }
+
+        h1 {
+            text-align: center;
+            color: #7EB6FF;
+            font-size: 2.5em;
+            margin-bottom: 40px;
+            text-shadow: 0 0 10px rgba(126, 182, 255, 0.5);
+        }
+
+        .section {
+            background: rgba(255, 255, 255, 0.1);
+            backdrop-filter: blur(10px);
+            border: 1px solid rgba(255, 255, 255, 0.2);
+            padding: 25px;
+            margin: 30px 0;
+            border-radius: 15px;
+            box-shadow: 0 8px 32px 0 rgba(31, 38, 135, 0.37);
+            transition: transform 0.3s ease;
+            text-align: center;
+        }
+
+        .section:hover {
+            transform: translateY(-5px);
+        }
+
+        h2 {
+            color: #7EB6FF;
+            margin-bottom: 20px;
+            font-size: 1.5em;
+        }
+
+        .button {
+            background: linear-gradient(45deg, #3498db, #2980b9);
+            color: white;
+            padding: 12px 24px;
+            border: none;
+            border-radius: 25px;
+            cursor: pointer;
+            margin: 5px;
+            transition: all 0.3s ease;
+            text-transform: uppercase;
+            letter-spacing: 1px;
+            font-weight: bold;
+        }
+
+        .button:hover {
+            background: linear-gradient(45deg, #2980b9, #3498db);
+            transform: scale(1.05);
+            box-shadow: 0 0 15px rgba(52, 152, 219, 0.5);
+        }
+
+        .delete-button {
+            background: linear-gradient(45deg, #e74c3c, #c0392b);
+        }
+
+        .delete-button:hover {
+            background: linear-gradient(45deg, #c0392b, #e74c3c);
+        }
+
+        input[type="text"], input[type="file"] {
+            background: rgba(255, 255, 255, 0.1);
+            border: 1px solid rgba(255, 255, 255, 0.2);
+            padding: 10px 15px;
+            border-radius: 25px;
+            color: white;
+            width: 100%;
+            margin-bottom: 15px;
+        }
+
+        input[type="file"] {
+            padding: 10px;
+        }
+
+        #response, #getResponse, #uploadResponse, #deleteResponse {
+            background: rgba(0, 0, 0, 0.3);
+            padding: 15px;
+            margin-top: 15px;
+            border-radius: 10px;
+            border: 1px solid rgba(255, 255, 255, 0.1);
+        }
+
+        a {
+            color: #7EB6FF;
+            text-decoration: none;
+            transition: all 0.3s ease;
+        }
+
+        a:hover {
+            color: #ffffff;
+            text-shadow: 0 0 10px rgba(126, 182, 255, 0.5);
+        }
+
+        .astronaut {
+            position: fixed;
+            bottom: 20px;
+            right: 20px;
+            width: 100px;
+            height: 100px;
+            animation: float 6s ease-in-out infinite;
+        }
+
+        @keyframes float {
+            0% { transform: translateY(0px) rotate(0deg); }
+            50% { transform: translateY(-20px) rotate(5deg); }
+            100% { transform: translateY(0px) rotate(0deg); }
+        }
+    </style>
+</head>
+<body>
+    <div class="container">
+        <div class="section">
+            <h2>Session test 🚀</h2>
+            <form id="sessionForm">
+                <input type="text" id="nameInput" name="name" placeholder="input your ID" required>
+                <button type="submit" class="button">log in</button>
+            </form>
+        </div>
+    </div>
+
+    <div class="astronaut"></div>
+
+    <script>
+        document.getElementById('sessionForm').onsubmit = async (e) => {
+            e.preventDefault();
+            
+            const formData = new FormData();
+            const name = document.getElementById('nameInput').value;
+            formData.append('name', name);
+
+            try {
+                const response = await fetch("/cookie/session.py", {
+                    method: 'POST',
+                    body: formData
+                });
+
+                const result = await response.text();
+                window.location.href = '/cookie';
+                document.getElementById('uploadResponse').innerHTML = `Session Started for: ${result}`;
+            } catch (error) {
+                document.getElementById('uploadResponse').innerHTML = `Error: ${error.message}`;
+            }
+        };
+    </script>
+</body>
+</html>
+"""
+if not file_path:
+    print(html_content)
+    sys.exit(0)
+
+
+
+session_data = {}
+
+with open(file_path, "r") as file:
+    for line in file:
+        if "=" in line:
+            key, value = line.strip().split("=", 1)
+            session_data[key] = value
+
+html_content = f"""<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>42 Webserv Space Station</title>
+    <style>
+        * {{
+            margin: 0;
+            padding: 0;
+            box-sizing: border-box;
+        }}
+
+        body {{
+            font-family: 'Arial', sans-serif;
+            background: linear-gradient(to bottom, #0B1026 0%, #1B2735 50%, #090A0F 100%);
+            color: #fff;
+            min-height: 100vh;
+            padding: 20px;
+            position: relative;
+            overflow-x: hidden;
+        }}
+
+        body::before {{
+            content: '';
+            position: absolute;
+            width: 100%;
+            height: 100%;
+            top: 0;
+            left: 0;
+            background-image: 
+                radial-gradient(2px 2px at 20px 30px, #ffffff, rgba(0,0,0,0)),
+                radial-gradient(2px 2px at 40px 70px, #ffffff, rgba(0,0,0,0)),
+                radial-gradient(2px 2px at 50px 160px, #ffffff, rgba(0,0,0,0)),
+                radial-gradient(2px 2px at 90px 40px, #ffffff, rgba(0,0,0,0)),
+                radial-gradient(2px 2px at 130px 80px, #ffffff, rgba(0,0,0,0));
+            background-repeat: repeat;
+            animation: twinkle 5s infinite linear;
+            z-index: -1;
+        }}
+
+        .button {{
+            background: linear-gradient(45deg, #3498db, #2980b9);
+            color: white;
+            padding: 12px 24px;
+            border: none;
+            border-radius: 25px;
+            cursor: pointer;
+            margin: 5px;
+            transition: all 0.3s ease;
+            text-transform: uppercase;
+            letter-spacing: 1px;
+            font-weight: bold;
+        }}
+
+        .button:hover {{
+            background: linear-gradient(45deg, #2980b9, #3498db);
+            transform: scale(1.05);
+            box-shadow: 0 0 15px rgba(52, 152, 219, 0.5);
+        }}
+
+        @keyframes twinkle {{
+            from {{ transform: translateY(0); }}
+            to {{ transform: translateY(-100px); }}
+        }}
+
+        .container {{
+            max-width: 800px;
+            margin: 0 auto;
+            position: relative;
+        }}
+
+        p {{
+            margin-bottom: 10px;
+        }}
+
+        h1 {{
+            text-align: center;
+            color: #7EB6FF;
+            font-size: 2.5em;
+            margin-bottom: 40px;
+            text-shadow: 0 0 10px rgba(126, 182, 255, 0.5);
+        }}
+
+        .section {{
+            background: rgba(255, 255, 255, 0.1);
+            backdrop-filter: blur(10px);
+            border: 1px solid rgba(255, 255, 255, 0.2);
+            padding: 25px;
+            margin: 30px 0;
+            border-radius: 15px;
+            box-shadow: 0 8px 32px 0 rgba(31, 38, 135, 0.37);
+            transition: transform 0.3s ease;
+            text-align: center;
+        }}
+
+        .section:hover {{
+            transform: translateY(-5px);
+        }}
+
+        h3 {{
+            color: #7EB6FF;
+            margin-bottom: 30px;
+            font-size: 1.5em;
+            text-align: center;
+        }}
+
+        h2 {{
+            color: #13dd71;
+            margin: 50px;
+            font-size: 1.5em;
+            text-align: center;
+        }}
+
+        span {{
+            color: #13dd71;
+            margin: 20px;
+            font-size: 1em;
+            text-align: center;
+        }}
+
+        #response, #getResponse, #uploadResponse, #deleteResponse {{
+            background: rgba(0, 0, 0, 0.3);
+            padding: 15px;
+            margin-top: 15px;
+            border-radius: 10px;
+            border: 1px solid rgba(255, 255, 255, 0.1);
+        }}
+
+        a {{
+            color: #0c0737;
+            text-decoration: none;
+            transition: all 0.3s ease;
+        }}
+
+        a:hover {{
+            color: #ffffff;
+            text-shadow: 0 0 10px rgba(126, 182, 255, 0.5);
+        }}
+
+        @keyframes float {{
+            0% {{ transform: translateY(0px) rotate(0deg); }}
+            50% {{ transform: translateY(-20px) rotate(5deg); }}
+            100% {{ transform: translateY(0px) rotate(0deg); }}
+        }}
+    </style>
+</head>
+<body>
+    <div class="container">
+        <div class="section">
+            <h3>Welcome to Space🚀</h3>
+                <h2>{session_data['name']}</h2>
+            <span class="button"><a href="/" >Go Home</a></span>
+        </div>
+    </div>
+</body>
+</html>"""
+
+print(html_content)
+
+
