--- www/html/cookie/session/k6^XP-#QXOb5pV*K
@@ -1,1 +1,0 @@
-name=นิรุตติ์ สมรอด
